pyproject.toml
[tool.poetry]
name = "todoapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.119.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
asyncpg = "^0.30.0"
pydantic = "^2.12.2"
pydantic-settings = "^2.11.0"
alembic = "^1.17.0"
argon2-cffi = "^25.1.0"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
psycopg = "^3.2.13"
asyncio = "^4.0.0"
pytest = "^9.0.1"
httpx = "^0.28.1"
pytest-asyncio = "^1.3.0"
aiofiles = "^25.1.0"
jinja2 = "^3.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
